+++ just_audio/test/just_audio_test.dart
@@ -1769,6 +1769,7 @@ final icyMetadata = IcyMetadata(
   ),
   info: IcyInfo(
     title: 'title',
+    artist: 'artist',
     url: 'url',
   ),
 );
@@ -1784,6 +1785,7 @@ final icyMetadataMessage = IcyMetadataMessage(
   ),
   info: IcyInfoMessage(
     title: 'title',
+    artist: 'artist',
     url: 'url',
   ),
 );
@@ -2060,6 +2062,7 @@ class MockAudioPlayer extends AudioPlayerPlatform {
         ),
         info: IcyInfoMessage(
           title: 'title',
+          artist: 'artist',
           url: url,
         ),
       ),