+++ just_audio_platform_interface/CHANGELOG.md
@@ -22,7 +22,8 @@
 
 ## 4.2.1
 
-* Update minimum flutter version to 3.0.
+- Added support for HLS title and artist metadata.
+- Update minimum flutter version to 3.0.
 
 ## 4.2.0